tests/public_benchmark/ldbc/requirements_3.txt
neo4j-driver==1.4.0
pkg-resources==0.0.0
